Don't collapse the tree when selection changes
This is more intuitive since it prevents you losing your place
when moving around. Auto-expanding of tree still exists, which
remains useful, but the unhelpful behavior of collapsing on click
is no longer present.

I noticed that Gnome Builder has the behavior of collapsing when
clicking on folders. I think on there it makes more sense since
you usually do not need to select a folder to *edit* it. You just want
to collapse it or select its *children*.

But here, you often need to select a parent and it doesn't make sense
to close it when selecting it.